test deleteQuiz() passed

+++ QuieasyServer/src/actions/DeleteObjects.java
@@ -9,31 +9,31 @@ import org.hibernate.query.Query;
 import persistence.HibernateUtil;
 
 public class DeleteObjects {
-    public static Session session = HibernateUtil.getSessionFactory().openSession();
-    public static Message message = new Message();
+    public static Session session;
+    public static Message message;
 
-    public static Message deleteQuiz(Long id, String course)
+    public static Message deleteQuiz(String quizName, String courseName)
     {
         try {
+            session = HibernateUtil.getSessionFactory().openSession();
             session.beginTransaction();
-            Query queryCourse = session.getSession().createQuery("FROM Course WHERE courseName = :courseName ");
-            queryCourse.setParameter("courseName", course);
-            Course courseToAdd = (Course) queryCourse.list().get(0);
-            Long CourseId = courseToAdd.getId();
-            Query queryQuiz = session.getSession().createQuery("FROM Quiz WHERE quiz_Name= :id ");
-            queryQuiz.setParameter("id", id);
-
-            Quiz quizToDelete = (Quiz) queryQuiz.list().get(0);
-            session.delete(quizToDelete);
-            session.getTransaction().commit();
-
-            message.task = "QUIZ_DELETED";
-    }
-        catch(Exception e)
+            Course course = session.getSession().createQuery("FROM Course WHERE course_name = :course", Course.class).setParameter("course", courseName).getSingleResult();
+            Quiz quizToDelete = session.getSession().createQuery("FROM Quiz WHERE quiz_name= :quiz AND id_course = :course", Quiz.class)
+                    .setParameter("quiz", quizName).setParameter("course", course.getId()).getSingleResult();
+            message = new Message();
+            if(quizToDelete == null){
+                message.status = false;
+            } else {
+                session.delete(quizToDelete);
+                session.getTransaction().commit();
+                message.status = true;
+            }
+    } catch(Exception e)
         {
             // if the error message is "out of memory",
             // it probably means no database file is found
             System.err.println(e.getMessage());
+            message.status = false;
         }
         finally
         {
@@ -55,14 +55,17 @@ public class DeleteObjects {
     public static Message deleteQuestion(Long questionID)
     {
         try {
+            session = HibernateUtil.getSessionFactory().openSession();
             session.beginTransaction();
             Question questionToDelete = session.getSession().createQuery("FROM Question WHERE id = :id ", Question.class).setParameter("id", questionID).getSingleResult();
+            message = new Message();
+
             if (questionToDelete == null) {
-                message.task = "DELETE_FAILED";
+                message.status = false;
             }else {
                 session.delete(questionToDelete);
                 session.getTransaction().commit();
-                message.task = "DELETE_OK";
+                message.status = true;
             }
         }
         catch(Exception e)
@@ -70,6 +73,7 @@ public class DeleteObjects {
             // if the error message is "out of memory",
             // it probably means no database file is found
             System.err.println(e.getMessage());
+            message.status = false;
         }
         finally
         {
